+++ src/services/quizzes.js
@@ -32,20 +32,50 @@ const getQuiz = async (quizId) => {
     }
 }
 
-const deleteQuiz = async (quizId) => {
-    const result = await prisma.quiz.delete({
+const updateQuiz = async (userId, quizId, quizInfo) => {
+    const result = await prisma.quiz.findUnique({
         where: {
             id: quizId
         }
     });
 
     if (!result) {
-        throw new httpError(400, "No quiz with id " + quizId + ".");
+        throw new httpError(400, `No quiz with id ${quizId}.`);
+    } else if (result.userId !== userId) {
+        throw new httpError(403, `You are not authorized to update quiz with id ${quizId}.`);
+    } else {
+        await prisma.quiz.update({
+            where: {
+                id: quizId
+            },
+            data: quizInfo
+        });
+    }
+}
+
+const deleteQuiz = async (userId, quizId) => {
+    const result = await prisma.quiz.findUnique({
+        where: {
+            id: quizId
+        }
+    });
+
+    if (!result) {
+        throw new httpError(400, `No quiz with id ${quizId}.`);
+    } else if (result.userId != userId) {
+        throw new httpError(403, `You are not authorized to delete quiz with id ${quizId}.`);
+    } else {
+        await prisma.quiz.delete({
+            where: {
+                id: quizId
+            }
+        });
     }
 }
 
 export default {
     createQuiz,
     getQuiz,
+    updateQuiz,
     deleteQuiz
 };
